Moved RowManager tests into its own package
This makes it a lot easier to only run RowManager tests.

Also ignored 3 more tests that do an "export()" - those still cause a segfault on my Mac. Will un-ignore them once that segfault bug is fixed.
