+++ marklogic-client-api/src/test/java/com/marklogic/client/test/rows/RowManagerFromDocDescriptorsTest.java
@@ -1,4 +1,4 @@
-package com.marklogic.client.test;
+package com.marklogic.client.test.rows;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
@@ -18,6 +18,7 @@ import com.marklogic.client.io.DocumentMetadataHandle;
 import com.marklogic.client.io.Format;
 import com.marklogic.client.io.JacksonHandle;
 import com.marklogic.client.io.StringHandle;
+import com.marklogic.client.test.Common;
 
 public class RowManagerFromDocDescriptorsTest extends AbstractRowManagerTest {
