+++ marklogic-client-api/src/test/java/com/marklogic/client/test/rows/RowManagerJoinDocColsTest.java
@@ -1,4 +1,4 @@
-package com.marklogic.client.test;
+package com.marklogic.client.test.rows;
 
 import static org.junit.Assert.assertEquals;
